airflow-docker/requirements-airflow.txt
pandas>=2.2,<3
numpy>=2,<3
scikit-learn>=1.4,<2
scipy>=1.11,<2
psycopg2-binary>=2.9,<3
redis>=5,<6
matplotlib>=3.8,<3.9